refactor Friend class and rename it to User
